+++ auction/src/components/Shop/ProductCard/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProductCard';
