+++ auction/src/components/Shop/Shop.js
@@ -6,7 +6,7 @@ import queryString from 'query-string';
 import PageLayout from '../PageLayout'
 import Wrapper from '../Common/Wrapper';
 import Grid from '../Common/Grid';
-import SubcategoryList from '../Home/SubcategoryList';
+import SubcategoryList from './SubcategoryList';
 import config from '../../config';
 import ProductCard from './ProductCard';
 import Divider from '../Common/Divider';
