+++ auction/src/components/Shop/SubcategoryList/index.js
@@ -1,0 +1,1 @@
+export { default } from './SubcategoryList';
